+++ internal/provider/services.go
@@ -41,6 +41,7 @@ import (
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/digitaltwins"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/disks"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/dns"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/services/dnsprivateresolver"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/domainservices"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/elastic"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/eventgrid"
@@ -85,7 +86,6 @@ import (
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/postgres"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/powerbi"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/privatedns"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/privatednsresolver"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/purview"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/recoveryservices"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/redis"
@@ -132,6 +132,7 @@ func SupportedTypedServices() []sdk.TypedServiceRegistration {
 		dashboard.Registration{},
 		databricks.Registration{},
 		disks.Registration{},
+		dnsprivateresolver.Registration{},
 		domainservices.Registration{},
 		eventhub.Registration{},
 		fluidrelay.Registration{},
@@ -143,7 +144,6 @@ func SupportedTypedServices() []sdk.TypedServiceRegistration {
 		mssql.Registration{},
 		nginx.Registration{},
 		policy.Registration{},
-		privatednsresolver.Registration{},
 		recoveryservices.Registration{},
 		resource.Registration{},
 		sentinel.Registration{},